Alteração das pastas para seguir o padrão solicitado pelo professor. -SCA
